a-h-event: always compare results of realpath()
In search for the duplicates, abrt-handle-event tries to avoid
identification of the processed problem itself by comparison of file
system paths of the compared problems. But it compares a result of
realpath() as the first argument with a concatenation of 'Dump Location
Path' and entries of that path as the second argument.

abrt-handle-event assumes that 'Dump Location Path' is the real path but
this assumption is not valid.

Thanks to Michael Calmer <mc@suse.de>

Signed-off-by: Jakub Filak <jfilak@redhat.com>